+++ src/helper/loader-options.ts
@@ -1,10 +1,18 @@
 import { getOptions } from 'loader-utils';
 import { PLATFORM_MAP } from './config';
 
+export const ALL_PLATFORM = 'ALL';
+
+
 export function shouldUseLoader(this: any, defaultPlatforms: Array<string> = []) {
   const options = getOptions(this) || {};
   const { platforms = defaultPlatforms  } = options;
   const platform = process.env.UNI_PLATFORM || '';
+
+  if (platforms === ALL_PLATFORM
+    || (platforms as string[]).indexOf(ALL_PLATFORM) > -1) {
+    return true;
+  }
 
   return (platforms  as string[]).includes(platform);
 }