+++ src/loader/cross-game-style/config.ts
@@ -1,0 +1,1 @@
+export const TIP_STYLE_NAME = '@TIP_STYLE_NAME';