+++ src/loader/inject-dynamic-style-mp/index.ts
@@ -1,0 +1,1 @@
+export { injectDynamicStyleMp as default } from './inject-dynamic-style-mp';