+++ src/loader/inject-dynamic-style-web/index.ts
@@ -1,0 +1,6 @@
+export {
+  getStyleList,
+} from './helper';
+export {
+  injectDynamicStyleWeb as default,
+} from './inject-dynamic-style-web';