+++ src/loader/replace-content/index.ts
@@ -1,0 +1,1 @@
+export { replaceContent as default } from './replace-content';