+++ src/loader/vue-directive/index.ts
@@ -1,0 +1,1 @@
+export { vueDirectiveLoader as default } from './vue-directive';